number of iterations now recorded as an int in MinuitFitterResult

+++ include/lsst/meas/multifit/MinuitFitter.h
@@ -41,7 +41,7 @@ public:
 
     int flags;
     double chisq;
-    double nIterations;
+    int nIterations;
     std::vector<double> parameters;
     Model::ConstPtr model;
     Eigen::MatrixXd covariance;
